unknown-aware "IF" and "NOT"
Removed parser-level booleans / nil; these are now
just special builtin bindings.

Removed parser-level #any-of.

This was problematic because it's hard to understand
whether #any-of(true false) means "either of the two
symbols 'true or 'false" or "either of the boolean values
true or false". With this change, all code should be
clear and unambiguous, even though some printable
representations are ambiguous.