implement (= 2 2.0); proper eq cmp for numerics
Featuring a really horrible hack for import-cycle
breaking, which has to be reconsidered later.

Currently the cyclebreaker hack makes it "dangerous"
to do things in init functions, before the cycles
are properly set up. (See e.g. how the initialisation
of MaybeValue was changed to accommodate this.)

Here's the particular cycle that was hard to break:

Equals(): fundamentally a symmetric function so we
          don't want to make it an interface function
          that needs to be re-implemented everywhere.
          Depends on "numerics" to do numeric equality
          comparison.
numerics: Depends on "anyof" for MaybeValue, because
          the result of some comparisons is Maybe.
anyof:    Depends on Equals() because lists of values
          need to be de-duplicated when a new AnyOf
          value is created.