recast any? and all? as unknown-aware folds
This makes most of the lists-of-unknown-length
aspirational testcase pass!

Split the "unclear" parts into their own testcase
which is still failing.

The "questionable" parts are, essentially:

   (length unknown) ==> ?
   (any? my-predicate unknown) ==> ?

I think the answers should be:
   (length unknown) ==> #number-in-range([0,inf) integer)
   (any? my-predicate unknown) ==> maybe

Note that both of these are "optimistic" --
they state what the value of the expression might
be, assuming that it does have a value (not an
error).