change sort implementation
I believed this one would be faster, but in fact
it's much slower. However, I think that's due to
a bug.. something with macroexpansions happening
over and over instead of once.

Now only due to quasiquote, and it happens a
lot because quasiquotes are used in the
implementation of cond. needs to be debugged.

Intuitively it feels like the new sort impl
should be faster once the bugs are fixed.